+++ src/main/java/me/minkh/app/exception/advice/ControllerAdvice.java
@@ -1,0 +1,19 @@
+package me.minkh.app.exception.advice;
+
+import me.minkh.app.exception.ErrorResponse;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+@RestControllerAdvice
+public class ControllerAdvice {
+
+    @ExceptionHandler(IllegalArgumentException.class)
+    public ResponseEntity<ErrorResponse> illegalArgumentExceptionHandler(IllegalArgumentException e) {
+        HttpStatus status = HttpStatus.BAD_REQUEST;
+        ErrorResponse errorResponse = new ErrorResponse(status.value(), e.getMessage());
+        return new ResponseEntity<>(errorResponse, status);
+    }
+
+}
